md-p-11_Crew-AI-and-RAG/AIssistant/pyproject.toml
[tool.poetry]
name = "aissistant"
version = "0.1.0"
description = ""
authors = ["Phenixjj <58008493+Phenixjj@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"
crewai = "^0.30.11"
python-dotenv = "^1.0.1"
tools = "^0.1.9"
llama-index = "^0.10.37"
pypdf = "^4.2.0"
pandas = "^2.2.2"
llama-index-experimental = "^0.1.3"
llama-index-llms-ollama = "^0.1.4"
llama-index-embeddings-ollama = "^0.1.2"
llama-index-llms-openai = "^0.1.20"
llama-index-core = "^0.10.37.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
